requirements.txt
torch
torchvision
pillow
opencv-python
numpy
tesseract
easyocr
matplotlib
pandas
streamlit